refactor: move image-list modules to image-picker
